+++ CHANGELOG.md
@@ -7,7 +7,9 @@ All notable changes to this project will be documented in this file.
 ## Unreleased
 
 ### Added
-- Initiative tracker now shows the active actor and has a next turn button
+- Initiative tracker update
+    - now shows the active actor and has a next turn button
+    - shows the current round number
 
 ## [0.8] - 2018-06-19